+++ app/components/SharedBoards.vue
@@ -12,6 +12,9 @@
                 :name="item.name"
                 :style="item.style"
                 :image="item.image"
+                :members="item.members"
+                :member-count="item.memberCount"
+                :unread-count="item.unreadCount"
             />
         </div>
     </div>